requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4
pydantic-settings>=2.1.0
pillow==10.1.0
opencv-python>=4.8.0
spacy==3.7.2
numpy<2.0.0
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
esdk-obs-python==3.23.9.1
streamlit==1.29.0
langchain==0.3.27
langchain-openai==0.3.33
langchain-core==0.3.76
openai==1.107.1